+++ src/server/controller-resolver.ts
@@ -15,7 +15,7 @@ import {IControllerMetadata} from "../interfaces/icontroller";
 import {IConnection} from "../interfaces/iconnection";
 import {IAction} from "../interfaces/iaction";
 import {IParam} from "../interfaces/iparam";
-import {Redirect} from "./redirect";
+import {StatusCode} from "./status-code";
 /**
  * Cookie parse regex
  * @type {RegExp}
@@ -315,7 +315,7 @@ export class ControllerResolver {
     }
     // check if action is present
     if (!isPresent(mappedAction)) {
-      throw new HttpError(500, `@${name}("${actionName}") is not defined on controller ${Metadata.getName(controllerProvider.provide)}`, {
+      throw new HttpError(StatusCode.Bad_Request, `@${name}("${actionName}") is not defined on controller ${Metadata.getName(controllerProvider.provide)}`, {
         name,
         actionName,
         resolvedRoute: this.resolvedRoute
@@ -857,7 +857,7 @@ export class Request {
    * @description
    * Set status code
    */
-  setStatusCode(code: number) {
+  setStatusCode(code: StatusCode) {
     this.controllerResolver.getEventEmitter().emit("statusCode", code);
   }
 
@@ -880,7 +880,7 @@ export class Request {
    * @description
    * Stops action chain
    */
-  redirectTo(url: string, code: Redirect) {
+  redirectTo(url: string, code: StatusCode) {
     this.stopChain();
     this.controllerResolver.getEventEmitter().emit("redirectTo", {
       url, code