+++ src/server/fake-http.ts
@@ -6,6 +6,7 @@ import {Inject} from "../decorators/inject";
 import {Readable, Writable} from "stream";
 import {Socket} from "net";
 import {ServerResponse, IncomingMessage} from "http";
+import {isObject} from "../core";
 
 export interface IFakeServerConfig {
 }
@@ -33,8 +34,21 @@ export function fakeHttpServer(Class: Function, config?: IFakeServerConfig): Fak
   ]);
   return fakeServerInjector.get(FakeServerApi);
 }
-
-
+/**
+ * @since 1.0.0
+ * @class
+ * @name FakeResponseApi
+ * @constructor
+ * @description
+ * FakeResponseApi api
+ *
+ * @private
+ */
+export interface FakeResponseApi {
+  getStatusCode(): number;
+  getBody(): string | Buffer;
+  getHeaders(): any;
+}
 /**
  * @since 1.0.0
  * @class
@@ -68,7 +82,7 @@ export class FakeServerApi {
    * @description
    * Fire GET Method
    */
-  GET(url: string, headers?: Object) {
+  GET(url: string, headers?: Object): Promise<FakeResponseApi> {
     let request = new FakeIncomingMessage();
     request.method = "GET";
     request.url = url;
@@ -83,7 +97,7 @@ export class FakeServerApi {
    * @description
    * Fire OPTIONS Method
    */
-  OPTIONS(url: string, headers?: Object) {
+  OPTIONS(url: string, headers?: Object): Promise<FakeResponseApi> {
     let request = new FakeIncomingMessage();
     request.method = "OPTIONS";
     request.url = url;
@@ -98,7 +112,7 @@ export class FakeServerApi {
    * @description
    * Fire HEAD Method
    */
-  HEAD(url: string, headers?: Object) {
+  HEAD(url: string, headers?: Object): Promise<FakeResponseApi> {
     let request = new FakeIncomingMessage();
     request.method = "HEAD";
     request.url = url;
@@ -113,7 +127,7 @@ export class FakeServerApi {
    * @description
    * Fire DELETE Method
    */
-  DELETE(url: string, headers?: Object) {
+  DELETE(url: string, headers?: Object): Promise<FakeResponseApi> {
     let request = new FakeIncomingMessage();
     request.method = "DELETE";
     request.url = url;
@@ -128,7 +142,7 @@ export class FakeServerApi {
    * @description
    * Fire TRACE Method
    */
-  TRACE(url: string, headers?: Object) {
+  TRACE(url: string, headers?: Object): Promise<FakeResponseApi> {
     let request = new FakeIncomingMessage();
     request.method = "TRACE";
     request.url = url;
@@ -143,7 +157,7 @@ export class FakeServerApi {
    * @description
    * Fire CONNECT Method
    */
-  CONNECT(url: string, headers?: Object) {
+  CONNECT(url: string, headers?: Object): Promise<FakeResponseApi> {
     let request = new FakeIncomingMessage();
     request.method = "CONNECT";
     request.url = url;
@@ -158,7 +172,7 @@ export class FakeServerApi {
    * @description
    * Fire POST Method
    */
-  POST(url: string, data?: string|Buffer, headers?: Object) {
+  POST(url: string, data?: string|Buffer, headers?: Object): Promise<FakeResponseApi> {
     let request = new FakeIncomingMessage();
     request.method = "POST";
     request.url = url;
@@ -178,7 +192,7 @@ export class FakeServerApi {
    * @description
    * Fire PUT Method
    */
-  PUT(url: string, data?: string|Buffer, headers?: Object) {
+  PUT(url: string, data?: string|Buffer, headers?: Object): Promise<FakeResponseApi> {
     let request = new FakeIncomingMessage();
     request.method = "PUT";
     request.url = url;
@@ -198,7 +212,7 @@ export class FakeServerApi {
    * @description
    * Fire PATCH Method
    */
-  PATCH(url: string, data?: string|Buffer, headers?: Object) {
+  PATCH(url: string, data?: string|Buffer, headers?: Object): Promise<FakeResponseApi> {
     let request = new FakeIncomingMessage();
     request.method = "PATCH";
     request.url = url;
@@ -219,8 +233,14 @@ export class FakeServerApi {
    * @description
    * Fire request
    */
-  private request(request: FakeIncomingMessage, response: FakeServerResponse) {
-    return fireRequest(this.getModules(), request, response);
+  private request(request: FakeIncomingMessage, response: FakeServerResponse): Promise<FakeResponseApi> {
+    return fireRequest(this.getModules(), request, response).then(data => {
+      return {
+        getStatusCode: () => response.statusCode,
+        getHeaders: () => response.getHeaders(),
+        getBody: () => response.getBody()
+      };
+    });
   }
 }
 
@@ -251,11 +271,15 @@ class FakeIncomingMessage extends Readable implements IncomingMessage {
   statusMessage?: string;
   socket: Socket;
 
-  protected _read(size: number): void {};
+  protected _read(size: number): void {
+  };
 
-  setTimeout(msecs: number, callback: Function): NodeJS.Timer { return null; };
+  setTimeout(msecs: number, callback: Function): NodeJS.Timer {
+    return null;
+  };
 
-  destroy(error?: Error) {};
+  destroy(error?: Error) {
+  };
 }
 
 
@@ -274,36 +298,160 @@ class FakeServerResponse extends Writable implements ServerResponse {
 
   // Extended base methods
 
-  statusCode: number;
+  statusCode: number = 200;
+  headers: any = {};
   statusMessage: string;
-  headersSent: boolean;
+  headersSent: boolean = false;
   sendDate: boolean;
-  finished: boolean;
+  message: string | Buffer;
+  finished: boolean = false;
 
-  writeContinue() {}
-
-
-  addTrailers(headers: any) {}
-
-  end() {
-    this.emit("finish");
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#writeContinue
+   * @private
+   * @description
+   * Write continue chunk
+   */
+  writeContinue() {
   }
 
-  writeHead(statusCode: number, headers?: any) {}
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#addTrailers
+   * @private
+   * @description
+   * Add Header Trailers
+   */
+  addTrailers(headers: any) {
 
+  }
 
-  setHeader(name: string, value: string | string[]) {};
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#end
+   * @private
+   * @description
+   * End request
+   */
+  end() {
+    this.emit("finish");
+    this.finished = true;
+  }
 
-  setTimeout(msecs: number, callback: Function): ServerResponse { return null; };
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#writeHead
+   * @private
+   * @description
+   * Write head
+   */
+  writeHead(statusCode: number, headers?: any) {
+    this.statusCode = statusCode;
+    if (isObject(headers)) {
+      Object.assign(this.headers, headers);
+    }
+    this.headersSent = true;
+  }
 
-  getHeader(name: string): string { return null; }
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#setHeader
+   * @private
+   * @description
+   * Set response header
+   */
+  setHeader(name: string, value: string | string[]) {
+    this.headers[name] = value;
+  };
 
-  removeHeader(name: string) { return null; }
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#getHeader
+   * @private
+   * @description
+   * Get response header
+   */
+  getHeader(name: string): string {
+    if (this.headers.hasOwnProperty(name)) {
+      return this.headers[name];
+    }
+    return null;
+  }
 
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#removeHeader
+   * @private
+   * @description
+   * Remove response header
+   */
+  removeHeader(name: string) {
+    if (this.headers.hasOwnProperty(name)) {
+      delete this.headers[name];
+    }
+  }
+
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#getHeaders
+   * @private
+   * @description
+   * Get headers
+   */
+  getHeaders(): {[key: string]: string | string[]} {
+    return this.headers;
+  }
+
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#getBody
+   * @private
+   * @description
+   * Get writed body
+   */
+  getBody(): string | Buffer {
+    return this.message;
+  }
+
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#getBody
+   * @private
+   * @description
+   * Get writed body
+   */
+  getStatusCode(): number {
+    return this.statusCode;
+  }
+
+  /**
+   * @since 1.0.0
+   * @function
+   * @name FakeServerResponse#write
+   * @private
+   * @description
+   * Write data
+   */
   write(value: Buffer | string,
         encoding?: string | Function,
         cb?: Function | string): boolean {
+    this.message = value;
     return false;
   }
+
+  setTimeout(msecs: number, callback: Function): ServerResponse {
+    return null;
+  };
 
 }