+++ src/server/request-resolver.ts
@@ -15,7 +15,7 @@ import {getModule} from "./bootstrap";
 import {ControllerResolver} from "./controller-resolver";
 import {EventEmitter} from "events";
 import {Injector} from "../injector/injector";
-import {IRedirect, getRedirectCode} from "./redirect";
+import {IRedirect, getCodeByStatus, StatusCode} from "./status-code";
 import {clean} from "../logger/inspect";
 
 /**
@@ -116,7 +116,7 @@ export class RequestResolver implements IAfterConstruct {
    * ControllerResolver status code default 200
    */
   @Inject("statusCode", true)
-  private statusCode: number;
+  private statusCode: StatusCode;
 
   /**
    * @param {String} contentType
@@ -169,7 +169,7 @@ export class RequestResolver implements IAfterConstruct {
     // force HttpError to be thrown
     if (!(data instanceof HttpError)) {
       let _error: Error = data;
-      data = new HttpError(500, _error.message, {});
+      data = new HttpError(StatusCode.Bad_Request, _error.message, {});
       data.stack = _error.stack;
     }
     // log error message
@@ -210,7 +210,7 @@ export class RequestResolver implements IAfterConstruct {
             response: response,
             type: typeof response
           });
-          throw new HttpError(500, "ResponseType must be string or buffer", {
+          throw new HttpError(StatusCode.Bad_Request, "ResponseType must be string or buffer", {
             response
           });
         }
@@ -229,7 +229,7 @@ export class RequestResolver implements IAfterConstruct {
         break;
       case RenderType.REDIRECT:
         this.response.setHeader("Location", this.redirectTo.url);
-        this.response.writeHead(getRedirectCode(this.redirectTo.code));
+        this.response.writeHead(getCodeByStatus(this.redirectTo.code));
         this.response.end();
         break;
       default:
@@ -263,7 +263,7 @@ export class RequestResolver implements IAfterConstruct {
         return metadata.name === resolvedModule.controller;
       });
     if (!isPresent(controllerProvider)) {
-      throw new HttpError(500, `You must define controller within current route: ${resolvedModule.resolvedRoute.route}`, {
+      throw new HttpError(StatusCode.Bad_Request, `You must define controller within current route: ${resolvedModule.resolvedRoute.route}`, {
         controllerName: resolvedModule.controller,
         actionName: resolvedModule.action,
         resolvedRoute: resolvedModule.resolvedRoute